push 1

+++ problem/models.py
@@ -30,21 +30,25 @@ def printProblem(p):
     json_data = p.to_json()
     print(json_data)
 
+def generateProblemObjectFromRequest(p, req_data, mode):
+    p.question = req_data['question']
+    p.questionNo = req_data['questionNo']
+    p.description = req_data['description']
+    p.answer = req_data['answer']
+    p.note = req_data['note']
+    p.importance = req_data['importance']
+    p.category = req_data['category']
+    p.sub_category = req_data['sub_category']
+    if mode == 'add':
+        p.add_date = datetime.datetime.now()
+    p.modify_date = datetime.datetime.now()
+    return p
+
 def addProblem():
     try:
         req_data = request.get_json(force=True)
-        p = Problem()
-        p.description = req_data['description']
-        p.answer = req_data['answer']
-        p.hint = req_data['hint']
-        p.note = req_data['note']
-        p.importance = req_data['importance']
-        p.category = req_data['category']
-        p.sub_category = req_data['subCategory']
-        p.review_dates.append(req_data['lastReviewDate'])
-        p.add_date = datetime.datetime.now()
-        p.modify_date = datetime.datetime.now()
-        #printProblem(p)
+        p = Problem() 
+        generateProblemObjectFromRequest(p, req_data, 'add')
         p.save()
         return Response("OK", status=200)
     except Exception as e:
@@ -54,20 +58,9 @@ def editProblem(_id):
     try:
         req_data = request.get_json(force=True)
         p = Problem.objects.get(id=_id)
-        printProblem(p)
+        # printProblem(p)
         if p:
-            p.question = req_data['question']
-            p.questionNo = req_data['questionNo']
-            p.description = req_data['description']
-            p.answer = req_data['answer']
-            #p.hint = req_data['hint']
-            p.note = req_data['note']
-            p.importance = req_data['importance']
-            p.category = req_data['category']
-            p.sub_category = req_data['sub_category']
-            #p.review_dates.append(req_data['lastReviewDate'])
-            p.add_date = datetime.datetime.now()
-            p.modify_date = datetime.datetime.now()
+            generateProblemObjectFromRequest(p, req_data, 'edit')
             p.save()
             return Response("OK", status=200)
 
@@ -93,7 +86,11 @@ def getProblemById(_id):
 def deleteProblemById(_id):
     try:
         p = Problem.objects(id=_id).first()
-        p.delete()
-        return True
-    except:
-        return False
+        if p:
+            p.delete()
+            return Response("OK", status=200)
+
+        return Response("Problem not found", status=404)
+
+    except Exception as e:
+            return Response("Data is not updated " + str(e), status=400)
